backend/requirements.txt
# Core dependencies
fastapi
uvicorn
pydantic
starlette
python-dotenv
requests
httpx
websockets

# Database
sqlalchemy
alembic
aiosqlite

# Authentication
python-jose[cryptography]
passlib[bcrypt]
python-multipart
PyJWT

# Github API
PyGithub
gitingest

# Utilities
structlog
python-dateutil
ujson
pyyaml
Jinja2

# File handling
aiofiles

# Core framework
rich
tiktoken

# AI & LLM
openai
litellm
mcp

# Memory & Knowledge (Supabase + Qdrant hybrid architecture)
supabase
psycopg2-binary
qdrant-client

# Free embeddings (sentence-transformers instead of OpenAI)
sentence-transformers

# Data processing
pandas
matplotlib
numpy

cachetools
asyncpg
unstructured
markdown
docx
psycopg2-binary
